% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas1006_output.R
\docType{data}
\name{atlas1006_output}
\alias{atlas1006_output}
\title{Atlas1006 Output Data}
\format{
## `atlas1006_output`
A data frame with 115 rows and 5 columns:
\describe{
  \item{taxon}{Taxon name}
  \item{aldex2, ancombc, deseq2}{Whether each method found a significant DAA}
  \item{score}{Number of methods that found the taxon to be significant}
  ...
}
}
\source{
library(microbiome)
}
\usage{
atlas1006_output
}
\description{
Output from running Atlas1006 data through compare_DAA_methods()
}
\keyword{datasets}
