% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_methods.R
\name{compare_DAA_methods}
\alias{compare_DAA_methods}
\title{Compare methods of microbial differential abundance analysis (DAA)}
\usage{
compare_DAA_methods(ps, group, prev.thr = 0.1)
}
\arguments{
\item{ps}{a phyloseq object containing the data that is to be analyzed.}

\item{group}{a string name of a binary variable to compare.}

\item{prev.thr}{the prevalence threshold, between (0, 1]. Default value is 0.1}
}
\value{
A data frame containing whether each method found a statistically
significant difference between the groups for each taxon.
}
\description{
`compare_DAA_methods()` applies three popular methods of DAA to a phyloseq object
and returns a table with all three results
}
\examples{
library(microbiome)
data(atlas1006)
atlas1006 <- phyloseq::subset_samples(atlas1006, (bmi_group == "lean" | bmi_group == "obese"))
compare_DAA_methods(ps = atlas1006, group = "bmi_group", prev.thr = 0.1)

}
