% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{get_graph_input2}
\alias{get_graph_input2}
\title{Private Helper Function}
\usage{
get_graph_input2(daa_output)
}
\arguments{
\item{daa_ouput}{A dataframe to visualize.}
}
\description{
Process data into a suitable format for visualization with
visualize_overlap
}
