% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_result_heatmap.R
\name{plot_result_heatmap}
\alias{plot_result_heatmap}
\title{Visualize resulting p-values from differential abundance analysis (DAA) methods and meta-analysis using Fisher's method}
\usage{
plot_result_heatmap(daa_output)
}
\value{
An interactive heatmap containing the proportion of Taxa found significantly different
by the methods.
}
\description{
\code{plot_result_heatmap()} visualizes the -log(p-values) resulting from three
different methods of metagenomic DAA (DESeq2, ANCOM-BC, and ALDEx2), and their meta-analysis
using Fisher's method.
}
\details{
@param daa_output a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods() and germseq::atlas1006_output
}
\examples{
plot_result_heatmap(germseq::atlas1006_output)

}
\references{
Dewey, M. (2022). metap: Meta-Analysis of Significance Values. R package version 1.8.
\href{https://cran.r-project.org/web/packages/metap/index.html}{Link}.

Wickham, H. (2022). ggplot2: Create Elegant Data
Visualisations Using the Grammar of Graphics. R package version 3.4.0.
\href{https://cran.r-project.org/web/packages/ggplot2/index.html}{Link}.
}
