% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGermSeq.R
\name{runGermSeq}
\alias{runGermSeq}
\title{Launch Shiny app for germseq}
\usage{
runGermSeq()
}
\value{
Launches a Shiny page, no return value.
}
\description{
A function that launches the Shiny app for germseq.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
germseq::runTestingPackage()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}

Silva, A. (2022) TestingPackage: An Example R Package For
BCB410H. Unpublished. https://github.com/anjalisilva/TestingPackage
}
