% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{visualize_overlap}
\alias{visualize_overlap}
\title{Visualize overlap between differential abundance analysis (DAA) method results}
\usage{
visualize_overlap(daa_output)
}
\arguments{
\item{daa_output}{a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods()}
}
\value{
A pie chart containing the proportion of Taxa found significant by different
numbers of methods.
}
\description{
\code{visualize_overlap()} visualizes the overlap between taxon found significant
identified by three popular current methods of microbial differential abundance analysis,
DESeq2, ANCOM-BC, and ALDEx2.
}
\examples{
visualize_overlap(germseq::atlas1006_output)

}
