% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{visualize_overlap_pichart}
\alias{visualize_overlap_pichart}
\title{Visualize overlap between differential abundance analysis (DAA) method results}
\usage{
visualize_overlap_pichart(daa_output)
}
\arguments{
\item{daa_output}{a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods()}
}
\value{
A pichart containing the proportion of Taxa found significant by different
numbers of methods.
}
\description{
`visualize_overlap_pichart()` visualizes the overlap between taxon found significant
identified by three popular current methods of microbial differential abundance analysis,
DESeq2, ANCOM-BC, and ALDEx2.
}
\examples{
library(germseq)
data("atlas1006_output")
visualize_overlap_pichart(atlas1006_output)

}
