% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{visualize_performances}
\alias{visualize_performances}
\title{Visualize performances of different methods of microbial differential abundance analysis (DAA)}
\usage{
visualize_performances(daa_output)
}
\arguments{
\item{daa_output}{a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods() and germseq::atlas1006_output}
}
\value{
A stacked barchart containing the proportion of Taxa found significantly different
by the methods.
}
\description{
\code{visualize_performances()} visualizes the proportion of significant taxon
identified by three popular current methods of microbial differential abundance analysis,
DESeq2, ANCOM-BC, and ALDEx2.
}
\examples{
visualize_performances(germseq::atlas1006_output)

}
\references{
Wickham, H. (2022). ggplot2: Create Elegant Data
Visualisations Using the Grammar of Graphics. R package version 3.4.0.
\href{https://cran.r-project.org/web/packages/ggplot2/index.html}{Link}.
}
