% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{visualize_performances_barchart}
\alias{visualize_performances_barchart}
\title{Visualize performances of different methods of microbial differential abundance analysis (DAA)}
\usage{
visualize_performances_barchart(daa_output)
}
\arguments{
\item{daa_output}{a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods()}
}
\value{
A stacked barchart containing the proportion of Taxa found significantly different
by the methods.
}
\description{
`visualize_performances_barchart()` visualizes the proportion of significant taxon
identified by three popular current methods of microbial differential abundance analysis,
DESeq2, ANCOM-BC, and ALDEx2.
}
\examples{
library(germseq)
data("atlas1006_output")
visualize_performances_barchart(atlas1006_output)

}
