% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas1006_output.R
\docType{data}
\name{atlas1006_output}
\alias{atlas1006_output}
\title{Atlas1006 Output Data}
\format{
\subsection{\code{atlas1006_output}}{

A data frame with 115 rows and 6 columns:
\describe{
\item{taxon}{Taxon name}
\item{aldex2, ancombc, deseq2}{Benjamini-Hochberg adjusted p-value
resulting from the respective DAA analysis}
\item{fisher_p}{Resulting p-value from the Fisher method meta-analysis of the aldex2,
ancombc, and deseq2 adjusted p-values}
\item{rawcount}{Number of methods that found the taxon to be significant}
}
}
}
\source{
library(microbiome)
}
\usage{
atlas1006_output
}
\description{
Output from running Atlas1006 data through compare_DAA_methods()
}
\keyword{datasets}
