% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_method_comparison.R
\name{visualize_overlap}
\alias{visualize_overlap}
\title{Visualize overlap between differential abundance analysis (DAA) method results}
\usage{
visualize_overlap(daa_output)
}
\arguments{
\item{daa_output}{a data frame containing the results of the three methods, for format refer
to output from germseq::compare_DAA_methods()}
}
\value{
An interactive pie chart containing the proportion of Taxa found significant by different
numbers of methods.
}
\description{
\code{visualize_overlap()} visualizes the overlap between taxon found significant
identified by three popular current methods of microbial differential abundance analysis,
DESeq2, ANCOM-BC, and ALDEx2.
}
\examples{
visualize_overlap(germseq::atlas1006_output)

}
\references{
Dewey, M. (2022). metap: Meta-Analysis of Significance Values. R package version 1.8.
\href{https://cran.r-project.org/web/packages/metap/index.html}{Link}.

Wickham, H. (2022). ggplot2: Create Elegant Data
Visualisations Using the Grammar of Graphics. R package version 3.4.0.
\href{https://cran.r-project.org/web/packages/ggplot2/index.html}{Link}.
}
